+++ examples/asymmetric-retrieval.rs
@@ -30,14 +30,14 @@ use mongodb_voyageai::{Client, OutputDtype, model};
 /// Document with embedding from the large model.
 struct Document {
     title: String,
-    embedding: Vec<f64>,
+    embedding: Vec<f32>,
 }
 
 /// Computes cosine similarity between two vectors.
-fn cosine_similarity(a: &[f64], b: &[f64]) -> f64 {
-    let dot: f64 = a.iter().zip(b).map(|(x, y)| x * y).sum();
-    let norm_a: f64 = a.iter().map(|x| x * x).sum::<f64>().sqrt();
-    let norm_b: f64 = b.iter().map(|x| x * x).sum::<f64>().sqrt();
+fn cosine_similarity(a: &[f32], b: &[f32]) -> f32 {
+    let dot: f32 = a.iter().zip(b).map(|(x, y)| x * y).sum();
+    let norm_a: f32 = a.iter().map(|x| x * x).sum::<f32>().sqrt();
+    let norm_b: f32 = b.iter().map(|x| x * x).sum::<f32>().sqrt();
 
     if norm_a == 0.0 || norm_b == 0.0 {
         0.0
@@ -154,7 +154,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
         let query_vec = query_embedding.embedding(0).unwrap();
 
         // Find top 3 matches
-        let mut scored: Vec<(&Document, f64)> = documents
+        let mut scored: Vec<(&Document, f32)> = documents
             .iter()
             .map(|doc| {
                 let sim = cosine_similarity(query_vec, &doc.embedding);
